[KYUUBI #7379][2b/4][FOLLOWUP] Split ReactAgent into LlmStreamClient + composite MiddlewareDispatcher
ReactAgent had grown to mix three concerns: the ReAct control loop,
OpenAI streaming/chunk assembly, and middleware fold logic. Extract:

- LlmStreamClient: owns one streaming chat completion call, accumulates
  content + tool-call deltas, and exposes StreamResult.toAssistantMessage
  for SDK message construction. Depends only on the OpenAI SDK and
  AgentRunContext (emits ContentDelta via ctx.emit, no dispatcher
  reference).
- MiddlewareDispatcher: implements AgentMiddleware as a composite over
  the configured list. ReactAgent calls onAgentStart / onEvent /
  beforeLlmCall etc. on the composite the same way it would call any
  middleware; resolveApproval stays as a non-interface accessor for the
  approval flow's special case.

Also: afterToolCall now returns Decision<String> for symmetry with the
other interceptor hooks; ABORT marks ToolResult.isError=true so
listeners can distinguish a middleware-vetoed result from a successful
one. The emit-then-forward step splits cleanly: the composite runs
onEvent, and ReactAgent's ctx.setEventEmitter lambda forwards the
filtered event to the user's raw consumer.

ReactAgent's run() drops the eventConsumer parameter threading through
internal helpers — everywhere downstream uses ctx.emit().
